backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
boto3==1.35.36
anthropic>=0.40.0
python-dotenv==1.0.1
requests==2.32.3
